+++ packages/web3-eth/src/utils/detect_transaction_type.ts
@@ -48,5 +48,5 @@ export const detectTransactionType = (
 		transaction as unknown as Record<string, unknown>,
 	);
 
-export const detectRawTransactionType = (transaction: Buffer) =>
+export const detectRawTransactionType = (transaction: Uint8Array) =>
 	transaction[0] > 0x7f ? '0x0' : toHex(transaction[0]);