+++ packages/web3-validator/test/unit/validation/string.test.ts
@@ -14,8 +14,7 @@ GNU Lesser General Public License for more details.
 You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
-// eslint-disable-next-line import/no-extraneous-dependencies
-import { toBuffer } from 'web3-eth-accounts';
+import { hexToBytes } from 'ethereum-cryptography/utils';
 import {
 	isString,
 	isHex,
@@ -128,29 +127,29 @@ describe('validation', () => {
 			});
 			it('valid', () => {
 				const noLeadingZeroes = {
-					a: toBuffer('0x123'),
+					a: hexToBytes('0x1230'),
 				};
 				const noleadingZeroBytes = {
-					a: toBuffer('0x01'),
+					a: hexToBytes('0x01'),
 				};
-				const emptyBuffer = {
-					a: toBuffer('0x'),
+				const emptyUint8Array = {
+					a: hexToBytes('0x'),
 				};
 				const undefinedValue = {
 					a: undefined,
 				};
 
 				expect(() => validateNoLeadingZeroes(noLeadingZeroes)).not.toThrow();
-				expect(() => validateNoLeadingZeroes(emptyBuffer)).not.toThrow();
+				expect(() => validateNoLeadingZeroes(emptyUint8Array)).not.toThrow();
 				expect(() => validateNoLeadingZeroes(undefinedValue)).not.toThrow();
 				expect(() => validateNoLeadingZeroes(noleadingZeroBytes)).not.toThrow();
 			});
 			it('fails', () => {
 				const leadingZeroBytes = {
-					a: toBuffer('0x001'),
+					a: hexToBytes('0x0010'),
 				};
 				const onlyZeroes = {
-					a: toBuffer('0x0'),
+					a: hexToBytes('0x00'),
 				};
 
 				expect(() => validateNoLeadingZeroes(leadingZeroBytes)).toThrow();